Add classes to fetch jabber:iq:gateway prompt
This sends a request to the gateway to ask what strings to use in the UI
to describe the sort of data we are requesting from the user for later
JID translation.

Errors are ignored, and we just assume the gateway does not support
jabber:iq:gateway (since XEP-0100 does not require service discovery
support for historical reasons).