Actually submit JID for translation
If we got a prompt back earlier, then we know that the gateway supports
jabber:iq:gateway.  So we send off the user input, and then re-run our
form handler on whatever comes back.

Don't dismiss the dialog right away in JavaScript, but wait for the
server to be really done first.