Hash agility for Bits of Binary
Previously, code assumed bits of binary cids were sha1.  Now, if they are sha1
if uses the previous behaviour, but if they are another algoritm the algorithm
is included in the key.  Then 'sha1+' is only prepended to the key when
requesting if there is no hash specified already.

Fixes #1092