A SIMS message may contain a body along with the image, so show both
This already worked in the large-file case, but when there is a small file it
gets shown "as a sticker" which assumed there was no useful body on the message.
Now we put the body in as well.

The possibly-redundant URI body that some clients send is already handled by
existing code to clear out the paragraph, though that code could remove the
sticker preview entirely so we update it to remove only the offending link here.