Nice UI for adding a contact through a Gateway
Uses jabber:iq:gateway where supported (such as slidge and cheogram) and
otherwise falls back to jid escaping where supported, or finally traditional %
escaping.  When searching for contacts, automatically search all gateways that
are added to contacts as well, and show any relevant results.

Note: if the gateway returns an error because the format is bad, this doesn't
show the error since that doesn't really fit the search paradigm, and instead it
simply does not show a result for that gateway.