Add missing parts to EditorConfig file
I use tabs in my editor by default since I need >2 spaces to clearly
read indentation. This project luckily uses a value in that range, but I
have a conflict of spaces over tabs. If this `.editorconfig` file is to
exist, then it should more accurately reflect the rules in the project.