Extend suggested nginx websocket proxy timeouts.
With nginx running as a reverse proxy, movim frequently "flashes", displaying a toast, as its websocket is closed and reopened.  When this happens frequently the current chat log is often pulled down to the most recent message, making it hard to read a chat buffer backlog.

This most often occurs at times of low message traffic, such as when most of Europe is asleep.  Low message traffic reduces the frequency of websocket message flow, leading nginx to close the websocket due to proxy timeouts.

The default proxy timeout is sixty seconds.  Increasing the timeout leads to smoother backlog reading.